Refactor upgrade controller
